# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Multi-Merchant Social Commerce Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products (legacy)
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: platform
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        merchantId:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        currency:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
        isFeatured:
          type: boolean
        isSoldOut:
          type: boolean
        badge:
          type: string
        platform:
          type: string
          description: "Source platform: instagram, facebook, tiktok"
        postUrl:
          type: string
          description: "URL of the original social media post to embed"
        postedAt:
          type: string
        sellerUsername:
          type: string
        sellerAvatar:
          type: string
      required:
        - id
        - merchantId
        - title
        - description
        - price
        - currency
        - images
        - category
        - tags
        - likes
        - comments
        - shares
        - isFeatured
        - isSoldOut
        - platform
        - postedAt
        - sellerUsername
        - sellerAvatar
...